+++ packages/notifier/package.json
@@ -44,6 +44,7 @@
   },
   "files": [
     "src/",
+    "tools",
     "exported.js",
     "NEWS.md"
   ],
